perf: constraint subtree recompute, SIMD fix, scene cache, alloc elimination
MAR-087 through MAR-093: performance optimization phase.

Results (100 skeletons, 64 bones, 6 IK, 5 transform):
- constraint_us: 151 → 57μs (2.7x, subtree recompute + children map)
- render_us: 224 → 12μs (19x, PreparedScene cache with dirty flags)
- transform_us: 4μs (unified float SoA)
- Benchmark score: 15 → 34 (max 67 skeletons at 60fps)
- Zero heap allocations in constraint hot path

Fixes applied:
- Subtree-only world transform after IK/path/transform constraints
- SIMD copies SoA→AoS instead of recomputing via scalar
- Precomputed children adjacency map and bone tip cache
- PreparedScene cached between frames, bone palette memcpy only
- Persistent solved_poses and path constraint buffers
- Unified single-precision float transforms (72→24 bytes/bone)

+++ src/editor/shell_main.cpp
@@ -5681,18 +5681,18 @@ std::optional<ViewportLayout> build_viewport_layout(
         }
 
         has_active_bone = true;
-        min_x = std::min(min_x, world_transforms[bone_index].world_x);
-        min_y = std::min(min_y, world_transforms[bone_index].world_y);
-        max_x = std::max(max_x, world_transforms[bone_index].world_x);
-        max_y = std::max(max_y, world_transforms[bone_index].world_y);
+        min_x = std::min(min_x, static_cast<double>(world_transforms[bone_index].world_x));
+        min_y = std::min(min_y, static_cast<double>(world_transforms[bone_index].world_y));
+        max_x = std::max(max_x, static_cast<double>(world_transforms[bone_index].world_x));
+        max_y = std::max(max_y, static_cast<double>(world_transforms[bone_index].world_y));
     }
 
     if (!has_active_bone) {
         for (const auto& transform : world_transforms) {
-            min_x = std::min(min_x, transform.world_x);
-            min_y = std::min(min_y, transform.world_y);
-            max_x = std::max(max_x, transform.world_x);
-            max_y = std::max(max_y, transform.world_y);
+            min_x = std::min(min_x, static_cast<double>(transform.world_x));
+            min_y = std::min(min_y, static_cast<double>(transform.world_y));
+            max_x = std::max(max_x, static_cast<double>(transform.world_x));
+            max_y = std::max(max_y, static_cast<double>(transform.world_y));
         }
     }
